+++ CHANGELOG.md
@@ -7,9 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Sync all nominators and respective stake for all validators
-- Subscribe new session events to sync validators and nominators
-
 ## [0.1.0] - 2021-04-19
 
 - First release
@@ -84,3 +81,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fix dependencies
+
+## [0.4.2] - 2021-05-10
+
+### Changed
+
+- Sync all nominators and respective stake for all validators
+- Subscribe new session events to sync validators and nominators